requirements.txt
Django==2.1.7
django-filebrowser==3.11.1
django-grappelli==2.12.2
djangorestframework==3.9.2
gevent==1.4.0
ipython==7.4.0
PyMySQL==0.9.3
uWSGI==2.0.18